+++ hashicorp-terraform/modules/azure/standalone-linux-scaleset/providers.tf
@@ -1,0 +1,10 @@
+terraform {
+  experiments = [ module_variable_optional_attrs ]
+  
+  required_providers {
+    azurerm = {
+      source = "hashicorp/azurerm"
+      version = ">= 3.3.0"
+    }
+  }
+}